# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute Commerce Platform API
  description: |
    Enterprise-grade B2B FMCG commerce platform combining multi-channel ordering,
    embedded finance, and gig economy logistics.
  version: 1.0.0
  contact:
    name: OmniRoute Team
    email: api@omniroute.io
  license:
    name: Proprietary
servers:
  - url: https://api.omniroute.io/v1
    description: Production
  - url: https://staging-api.omniroute.io/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local Development
...
paths:
...
  /services:
    get:
      tags: [Services]
      summary: List service definitions
      operationId: listServices
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
    post:
      tags: [Services]
      summary: Create a service definition
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDefinition'
  /services/{serviceId}/execute:
    post:
      tags: [Services]
      summary: Execute a service
      operationId: executeService
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteServiceRequest'
      responses:
        '202':
          description: Service execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionResponse'
components:
...
  schemas:
...
    ServiceDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, published, deprecated]
    ServiceList:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceDefinition'
    CreateServiceRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
    ExecuteServiceRequest:
      type: object
      properties:
        input_data:
          type: object
    ExecutionResponse:
      type: object
      properties:
        execution_id:
          type: string
        status:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []